rename files and move files
