+++ src/main/scala/whu/edu/cn/debug/GWmodelUtil/SpatialRegression/LogisticRegression.scala
@@ -2,7 +2,7 @@ package whu.edu.cn.debug.GWmodelUtil.SpatialRegression
 
 import breeze.linalg.DenseVector
 
-object logistic_regression {
+object LogisticRegression {
 
   def logisticRegression(x: Array[DenseVector[Double]], y: DenseVector[Double]): Unit = {
