+++ src/main/scala/whu/edu/cn/debug/GWmodelUtil/SpatialRegression/PoissonRegression.scala
@@ -2,7 +2,7 @@ package whu.edu.cn.debug.GWmodelUtil.SpatialRegression
 
 import breeze.linalg.DenseVector
 
-object poisson_regression {
+object PoissonRegression {
 
   def poissonRegression(x: Array[DenseVector[Double]], y: DenseVector[Double]): Unit = {
