+++ src/main/scala/whu/edu/cn/debug/GWmodelUtil/Utils/OtherUtils.scala
@@ -1,8 +1,7 @@
 package whu.edu.cn.debug.GWmodelUtil.Utils
 
 import au.com.bytecode.opencsv._
-import breeze.linalg.{DenseVector, linspace}
-import breeze.plot._
+
 import org.apache.spark.SparkContext
 import org.apache.spark.mllib.stat.Statistics
 import org.apache.spark.rdd.RDD
@@ -14,7 +13,7 @@ import scala.collection.mutable.Map
 import scala.math.pow
 import scala.reflect.ClassTag
 
-object other_util {
+object OtherUtils {
 
   def readcsv(implicit sc: SparkContext, csvPath: String): RDD[Array[(String, Int)]] = {
     val data = sc.textFile(csvPath)
@@ -25,40 +24,25 @@ object other_util {
     csvdata.map(t => t.zipWithIndex)
   }
 
-  def timeSeries_acf(timeArr: Array[Double], timelag: Int = 20): Array[Double] = {
-    val acfarr = DenseVector.zeros[Double](timelag + 1).toArray
-    if (timelag > 0) {
-      val f = Figure()
-      val p = f.subplot(0)
-      for (i <- 0 until timelag + 1) {
-        acfarr(i) = getacf(timeArr, i)
-        val x = DenseVector.ones[Double](5) :*= i.toDouble
-        val y = linspace(0, acfarr(i), 5)
-        p += plot(x, y, colorcode = "[0,0,255]")
-      }
-      p.xlim = (-0.1, timelag + 0.1)
-      p.xlabel = "lag"
-      p.ylabel = "ACF"
-    } else {
-      throw new IllegalArgumentException("Illegal Argument of time lag")
-    }
-    acfarr
-  }
 
-  def getacf(timeArr: Array[Double], timelag: Int): Double = {
-    val lagArr = timeArr.drop(timelag)
-    val tarridx = timeArr.take(lagArr.length).zipWithIndex
-    val avg = timeArr.sum / timeArr.length
-    val tarr_avg = timeArr.map(t => t * avg).sum
-    val tarr_avg2 = tarridx.map(t => t._1 * avg).sum
-    val lag_avg = lagArr.map(t => t * avg).sum
-    //    var tarr_lag:Double =0.0
-    //    for(i<-0 until lagArr.length){
-    //      tarr_lag += timeArr(i)*lagArr(i)
-    //    }
-    val tarr_lag = tarridx.map(t => t._1 * lagArr(t._2)).sum
-    val acf = (tarr_lag - tarr_avg2 - lag_avg + pow(avg, 2) * lagArr.length) / (timeArr.map(t => t * t).sum - 2 * tarr_avg + pow(avg, 2) * timeArr.length)
-    acf
+  /**
+   * 输入要添加的属性数据和RDD，输出RDD
+   *
+   * @param sc           SparkContext
+   * @param shpRDD       源RDD
+   * @param writeArray   要写入的属性数据，Array形式
+   * @param propertyName 属性名，String类型，需要少于10个字符
+   * @return RDD
+   */
+  def writeRDD(sc: SparkContext, shpRDD: RDD[(String, (Geometry, Map[String, Any]))], writeArray: Array[Double], propertyName: String): RDD[(String, (Geometry, Map[String, Any]))] = {
+    if (propertyName.length >= 10) {
+      throw new IllegalArgumentException("the length of property name must not longer than 10!!")
+    }
+    val shpRDDidx = shpRDD.collect().zipWithIndex
+    shpRDDidx.map(t => {
+      t._1._2._2 += (propertyName -> writeArray(t._2))
+    })
+    sc.makeRDD(shpRDDidx.map(t => t._1))
   }
 
   def attributeSelectHead(csvRDD: RDD[Array[(String, Int)]], property: String): Array[String] = {
@@ -96,46 +80,9 @@ object other_util {
     resultArr
   }
 
-  /**
-   * 输入要添加的属性数据和RDD，输出RDD
-   *
-   * @param sc           SparkContext
-   * @param shpRDD       源RDD
-   * @param writeArray   要写入的属性数据，Array形式
-   * @param propertyName 属性名，String类型，需要少于10个字符
-   * @return RDD
-   */
-  def writeRDD(sc: SparkContext, shpRDD: RDD[(String, (Geometry, Map[String, Any]))], writeArray: Array[Double], propertyName: String): RDD[(String, (Geometry, Map[String, Any]))] = {
-    if (propertyName.length >= 10) {
-      throw new IllegalArgumentException("the length of property name must not longer than 10!!")
-    }
-    val shpRDDidx = shpRDD.collect().zipWithIndex
-    shpRDDidx.map(t => {
-      t._1._2._2 += (propertyName -> writeArray(t._2))
-    })
-    sc.makeRDD(shpRDDidx.map(t => t._1))
-  }
-
   def printArrArr[T: ClassTag](arrarr: Array[Array[T]]) = {
     val arrvec = arrarr.map(t => t.toVector)
     arrvec.foreach(println)
-  }
-
-  def testcorr(testshp: RDD[(String, (Geometry, Map[String, Any]))]): Unit = {
-    val time0: Long = System.currentTimeMillis()
-    val list1: List[Any] = Feature.get(testshp, "PURCHASE") //FLOORSZ,PROF
-    val list2: List[Any] = Feature.getNumber(testshp, "FLOORSZ")
-    //    println(list1)
-    //    println(list2)
-    val lst1: List[Double] = list1.collect({ case (i: String) => (i.toDouble) })
-    //    //val lst2:List[Double]=list2.collect({ case (i: String) => (i.toDouble) })
-    val lst2: List[Double] = list2.asInstanceOf[List[Double]]
-    val corr1 = corr2list(lst1, lst2)
-    val listtimeused = System.currentTimeMillis() - time0
-    println(s"time used is: $listtimeused")
-
-    val coor2 = corr2ml(testshp, "PURCHASE", "FLOORSZ")
-
   }
 
   def corr2list(lst1: List[Double], lst2: List[Double]): Double = {
